requirements-dev.txt
alembic==1.0.8
Click==7.0
cx-Oracle==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.0.2
Flask-Login==0.5.0
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.8
MarkupSafe==1.1.1
psycopg2==2.8.6
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==1.0.1
WTForms==2.3.1
